feat: Separate local vs platform deployment
