+++ cmd/meroxa/turbine_cli/golang/deploy.go
@@ -3,17 +3,11 @@ package turbineGo
 import (
 	"context"
 	"fmt"
-	"io"
 	"os"
 	"os/exec"
 	"path"
 	"regexp"
 
-	turbine "github.com/meroxa/turbine/deploy"
-
-	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/client"
-	"github.com/docker/docker/pkg/archive"
 	"github.com/meroxa/cli/cmd/meroxa/global"
 	"github.com/meroxa/cli/log"
 )
@@ -27,9 +21,14 @@ type Deploy struct {
 // runDeployApp runs the binary previously built with the `--deploy` flag which should create all necessary resources
 func runDeployApp(ctx context.Context, l log.Logger, appPath, appName, imageName string) error {
 	l.Infof(ctx, "Deploying application %q...", appName)
+	var cmd *exec.Cmd
 
-	// TODO: Check here if imageName is ""
-	cmd := exec.Command(appPath+"/"+appName, "--deploy", "--imagename", imageName) // nolint:gosec
+	if imageName != "" {
+		cmd = exec.Command(appPath+"/"+appName, "--deploy", "--imagename", imageName) // nolint:gosec
+	} else {
+		cmd = exec.Command(appPath+"/"+appName, "--deploy")
+	}
+
 	accessToken, refreshToken, err := global.GetUserToken()
 	if err != nil {
 		return err
@@ -45,55 +44,6 @@ func runDeployApp(ctx context.Context, l log.Logger, appPath, appName, imageName
 
 	l.Info(ctx, string(stdout))
 	l.Info(ctx, "deploy complete!")
-	return nil
-}
-
-func (*Deploy) buildImage(ctx context.Context, l log.Logger, pwd, imageName string) error {
-	l.Infof(ctx, "Building image %q located at %q", imageName, pwd)
-	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
-	if err != nil {
-		l.Errorf(ctx, "unable to init docker client; %s", err)
-	}
-
-	// Generate dockerfile
-	err = turbine.CreateDockerfile(pwd)
-	if err != nil {
-		return err
-	}
-
-	// Read local Dockerfile
-	tar, err := archive.TarWithOptions(pwd, &archive.TarOptions{
-		Compression:     archive.Uncompressed,
-		ExcludePatterns: []string{".git", "fixtures"},
-	})
-	if err != nil {
-		l.Errorf(ctx, "unable to create tar; %s", err)
-	}
-
-	buildOptions := types.ImageBuildOptions{
-		Context:    tar,
-		Dockerfile: "Dockerfile",
-		Remove:     true,
-		Tags:       []string{imageName}}
-
-	resp, err := cli.ImageBuild(
-		ctx,
-		tar,
-		buildOptions,
-	)
-	if err != nil {
-		l.Errorf(ctx, "unable to build docker image; %s", err)
-	}
-	defer func(Body io.ReadCloser) {
-		err = Body.Close()
-		if err != nil {
-			l.Errorf(ctx, "unable to close docker build response body; %s", err)
-		}
-	}(resp.Body)
-	_, err = io.Copy(os.Stdout, resp.Body)
-	if err != nil {
-		l.Errorf(ctx, "unable to read image build response; %s", err)
-	}
 	return nil
 }
 
@@ -131,30 +81,21 @@ func (gd *Deploy) Deploy(ctx context.Context, appPath string, l log.Logger) erro
 		return err
 	}
 
-	// to determine whether we need to call POST /sources and POST /build
+	// check for image instances
 	if ok, err := needsToBuild(appPath, appName); ok {
 		if err != nil {
 			l.Errorf(ctx, err.Error())
 			return err
 		}
 
-		//fqImageName will be eventually taken from the build endpoint
-		fqImageName = gd.DockerHubUserNameEnv + "/" + appName
-
-		err = gd.buildImage(ctx, l, appPath, fqImageName)
-		if err != nil {
-			l.Errorf(ctx, "unable to build image; %q\n%s", fqImageName, err)
-			return err
+		if gd.LocalDeployment {
+			fqImageName, err = gd.getDockerImageName(ctx, l, appPath, appName)
+			if err != nil {
+				return err
+			}
+		} else {
+			fqImageName, err = gd.getPlatformImage(ctx, l)
 		}
-
-		// this will go away when using the new service
-		err = gd.pushImage(l, fqImageName)
-		if err != nil {
-			l.Errorf(ctx, "unable to push image; %q\n%s", fqImageName, err)
-			return err
-		}
-
-		// TODO: Keep polling for building statsu until it's successful
 	}
 
 	// creates all resources
